requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0 